#9 add new sql_script and new class Investments

+++ src/main/java/com/cashmm/cashflow/user/User.java
@@ -14,7 +14,6 @@ import java.sql.Timestamp;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Builder
 @Data
@@ -26,7 +25,7 @@ public class User implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     private String firstname;
     private String lastname;
     private String email;
